requirements.txt
psycopg[binary]>=3.0
pandas>=2.0
python-dotenv>=1.0
python-telegram-bot>=20.0